deduplicate code in build tags that isn't using mutex in either
